investlib-greeks/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_system"

[project]
name = "investlib-greeks"
version = "0.3.0"
description = "MyInvest V0.3 - Options Greeks calculation with py_vollib"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "py_vollib_vectorized>=0.1.0",
    "scipy>=1.11.0"
]
